LPD-74587 Relocate model classes into the right package
