+++ modules/apps/frontend-data-set/frontend-data-set-admin-web/src/main/java/com/liferay/frontend/data/set/admin/web/internal/model/SystemFDSEntryInfo.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.frontend.data.set.admin.web.internal.frontend.data.set.model;
+package com.liferay.frontend.data.set.admin.web.internal.model;
 
 import com.liferay.frontend.data.set.SystemFDSEntry;
